+++ spreadsheet-api/src/main/java/ec/util/spreadsheet/Book.java
@@ -35,6 +35,7 @@ import java.util.function.ObjIntConsumer;
 import nbbrd.service.Mutability;
 import nbbrd.service.Quantifier;
 import nbbrd.service.ServiceDefinition;
+import nbbrd.service.ServiceSorter;
 import org.checkerframework.checker.index.qual.NonNegative;
 import org.checkerframework.checker.nullness.qual.NonNull;
 
@@ -155,6 +156,11 @@ public abstract class Book implements Closeable {
          */
         @NonNull
         abstract public String getName();
+
+        @ServiceSorter(reverse = true)
+        public int getRank() {
+            return UNKNOWN_RANK;
+        }
 
         //<editor-fold defaultstate="collapsed" desc="Loading methods">
         /**
@@ -300,6 +306,10 @@ public abstract class Book implements Closeable {
                     || Number.class.isAssignableFrom(type)
                     || String.class.isAssignableFrom(type);
         }
+
+        public static final int NATIVE_RANK = Byte.MAX_VALUE;
+        public static final int WRAPPED_RANK = 0;
+        public static final int UNKNOWN_RANK = -1;
     }
 
     private static IOException translate(FileNotFoundException ex) {